+++ serialize/src/test.rs
@@ -267,3 +267,30 @@ fn test_biguint() {
         .unwrap();
     assert_eq!(bytes, expected);
 }
+
+#[test]
+fn test_serialize_macro() {
+    // Make a bunch of distinctly typed values
+    let val1 = [vec![1u8, 2u8, 3u8], vec![4u8, 5u8, 6u8]]
+        .into_iter()
+        .collect::<BTreeSet<_>>();
+    let val2: core::marker::PhantomData<Dummy> = core::marker::PhantomData;
+    let val3 = Some(10u64);
+    let val4 = 192830918usize;
+    let val5 = [1u64, 2, 3, 4, 5].into_iter().collect::<LinkedList<_>>();
+
+    // Make sure the serialization macro matches just serializing them as a tuple
+    let mut tuple_bytes = Vec::new();
+    (
+        val1.clone(),
+        val2.clone(),
+        val3.clone(),
+        val4.clone(),
+        val5.clone(),
+    )
+        .serialize_uncompressed(&mut tuple_bytes)
+        .unwrap();
+    let macro_bytes = serialize_to_vec![val1, val2, val3, val4, val5].unwrap();
+
+    assert_eq!(tuple_bytes, macro_bytes);
+}